Changing in class propertry on behavior.
