+++ src/Dali/Dali/Windows/TransparentWindow.xaml
@@ -3,7 +3,16 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        ShowInTaskbar="False" Background="Transparent" Opacity="0.5"
-        mc:Ignorable="d" WindowState="Maximized" AllowsTransparency="True" WindowStyle="None">
+        xmlns:behaviors="clr-namespace:RedSharp.Dali.View.Behaviors"
+        xmlns:i="http://schemas.microsoft.com/xaml/behaviors" 
+        ShowInTaskbar="False" Background="Transparent" 
+        mc:Ignorable="d" WindowState="Maximized" 
+        Opacity="{Binding Opacity}"
+        AllowsTransparency="True" WindowStyle="None">
+
+    <i:Interaction.Behaviors>
+        <behaviors:InputTransparencyBehavior IsInputTransparent="{Binding IsTransparent}"/>
+    </i:Interaction.Behaviors>
+    
     <Image Source="{Binding Item.Image, Converter={StaticResource SharpImageToBitmapSourceConverter}}"/>
 </Window>
